requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
opencv-python>=4.5.0
tensorflow>=2.8.0
scikit-learn>=1.0.0
scikit-image>=0.18.0
streamlit>=1.10.0
pillow>=8.0.0
albumentations>=1.1.0